only author can remove and edit comments

+++ src/api/posts/deletecomment.php
@@ -9,15 +9,26 @@ header('Access-Control-Allow-Credentials: true');
 header('Content-Type: application/json');
 
 if (isset($_SESSION['user'])) {
-    if (isset($_POST['id']) && $_POST['user_id'] === $_SESSION['user_id']) {
+    if (isset($_POST['id'])) {
         $commentId = trim(filter_var($_POST['id'], FILTER_SANITIZE_STRING));
 
-        $query = "DELETE FROM comments WHERE id = :cid";
+        $query = "SELECT user_id FROM comments WHERE id = :cid";
         $statement = $pdo->prepare($query);
-        $statement->bindParam(':cid', $commentId, PDO::PARAM_STR);
+        $statement->bindParam(':cid', $commentId,   PDO::PARAM_STR);
         $statement->execute();
-        echo json_encode(array('message' => 'The post was updated'));
-        http_response_code(200);
+        $author = $statement->fetch(PDO::FETCH_ASSOC);
+        //      echo json_encode('hej');
+        // exit;
+
+        if ($author['user_id'] === $_SESSION['user']) {
+            $query = "DELETE FROM comments WHERE id = :cid AND user_id = :uid";
+            $statement = $pdo->prepare($query);
+            $statement->bindParam(':cid', $commentId, PDO::PARAM_STR);
+            $statement->bindParam(':uid', $author['user_id'], PDO::PARAM_STR);
+            $statement->execute();
+            echo json_encode(array('message' => 'The post was updated'));
+            http_response_code(200);
+        }
     }
 } else {
     echo json_encode(array('message' => 'Not logged in'));
